P4: Project Feedback/contact page
Create a feedback/contact page for your project. Creating the backend for this is beyond the scope of this class, so we will submit via email.

Create a form with at least three fields and a submit button. Use form field tags from what you learned.

Modify your form tag to submit the results to yourself via email.