P1: Project Navagation
Take your home page and do the following. If you used a pre-made template with navigation, you will need to modify it. This will be for the project only rename your home page home.html. This will differentiate the project from the files we have been creating. Open your home page and then save it with the name of each of your four other pages.